+++ .travis.yml
@@ -1,0 +1,13 @@
+language: python
+
+python:
+    - "2.7"
+
+install:
+    - python setup.py install
+
+script:
+    - cd tests && nosetests -v -s
+
+notifications:
+    email: false
